Welcome,I am Rev.Russell Britton
Some of you may of heard of me or have seen my work.Im now focusing on a collection of christian coins to help keep god in focus and teach the lessons of the bible. Im creating these coins to intrigue their viewers so they may research more of the story and possibly read the good book to learn of its knowledge. God bless.